Separate layout react context from layout components
