+++ assets/src/edit-story/app/layout/index.js
@@ -14,6 +14,5 @@
  * limitations under the License.
  */
 
-export { default } from './layout';
 export { default as useLayout } from './useLayout';
 export { default as LayoutProvider } from './layoutProvider';
